Changed data types to int32_t and added AA notification to resend data

+++ modules/J2534/j2534plugin.h
@@ -17,15 +17,15 @@ class J2534Plugin : public QObject, PluginInterface {
     Q_PLUGIN_METADATA(IID "org.aselafernando.headunit.j2534" FILE "config.json")
     Q_INTERFACES(PluginInterface)
 
-    Q_PROPERTY(double iat      MEMBER m_iat      NOTIFY iatUpdated)
-    Q_PROPERTY(double maf      MEMBER m_maf      NOTIFY mafUpdated)
-    Q_PROPERTY(double stftb1   MEMBER m_stftb1   NOTIFY stftb1Updated)
-    Q_PROPERTY(double stftb2   MEMBER m_stftb2   NOTIFY stftb2Updated)
-    Q_PROPERTY(double ect      MEMBER m_ect      NOTIFY ectUpdated)
-    Q_PROPERTY(int    vss      MEMBER m_vss      NOTIFY vssUpdated)
-    Q_PROPERTY(double rpm      MEMBER m_rpm      NOTIFY rpmUpdated)
-    Q_PROPERTY(double accPedal MEMBER m_accPedal NOTIFY accPedalUpdated)
-    Q_PROPERTY(int    gear     MEMBER m_gear     NOTIFY gearUpdated)
+    Q_PROPERTY(int32_t iat      MEMBER m_iat      NOTIFY iatUpdated)
+    Q_PROPERTY(double  maf      MEMBER m_maf      NOTIFY mafUpdated)
+    Q_PROPERTY(double  stftb1   MEMBER m_stftb1   NOTIFY stftb1Updated)
+    Q_PROPERTY(double  stftb2   MEMBER m_stftb2   NOTIFY stftb2Updated)
+    Q_PROPERTY(int32_t ect      MEMBER m_ect      NOTIFY ectUpdated)
+    Q_PROPERTY(int32_t vss      MEMBER m_vss      NOTIFY vssUpdated)
+    Q_PROPERTY(double  rpm      MEMBER m_rpm      NOTIFY rpmUpdated)
+    Q_PROPERTY(double  accPedal MEMBER m_accPedal NOTIFY accPedalUpdated)
+    Q_PROPERTY(int32_t gear     MEMBER m_gear     NOTIFY gearUpdated)
 
     //Q_PROPERTY(QVariantMap ports READ getPorts NOTIFY portsUpdated)
     //Q_PROPERTY(bool connected MEMBER m_connected NOTIFY connectedUpdated)
@@ -44,15 +44,15 @@ public:
 public slots:
     void eventMessage(QString id, QVariant message) override;
 
-    void handleIAT(const double&);
+    void handleIAT(const int32_t&);
     void handleMAF(const double&);
     void handleSTFTB1(const double&);
     void handleSTFTB2(const double&);
-    void handleECT(const double&);
-    void handleVSS(const int&);
+    void handleECT(const int32_t&);
+    void handleVSS(const int32_t&);
     void handleRPM(const double&);
     void handleACCPEDAL(const double&);
-    void handleGEAR(const int&);
+    void handleGEAR(const int32_t&);
 
 signals:
     void message(QString id, QVariant message);
@@ -74,16 +74,16 @@ private slots:
     void settingsChanged(const QString &key, const QVariant &value);
 
 private:
-    double m_iat = 0;
+    int32_t m_iat = 0;
     double m_maf = 0;
     double m_stftb1 = 0;
     double m_stftb2 = 0;
-    double m_ect = 0;
-    int m_vss = 0;
+    int32_t m_ect = 0;
+    int32_t m_vss = 0;
     double m_rpm = 0;
     double m_accPedal = 0;
-    int m_gear = 0;
-    int m_reversePageIndex = 3;
+    int32_t m_gear = 0;
+    uint32_t m_reversePageIndex = 3;
 
     void startWorker();
     void stopWorker();
@@ -109,15 +109,15 @@ public slots:
     void j2534Connect();
 
 signals:
-    void iat(const double& result);
+    void iat(const int32_t& result);
     void maf(const double& result);
     void stftb1(const double& result);
     void stftb2(const double& result);
-    void ect(const double& result);
+    void ect(const int32_t& result);
     void accPedal(const double& result);
     void rpm(const double& result);
-    void vss(const int& result);
-    void gear(const int& result);
+    void vss(const int32_t& result);
+    void gear(const int32_t& result);
 };
